// bd-client/src/app/models/enums/color.ts
export enum Color {
  White = '#ffffff',
  Background = '#f9f7fe',
  Light = '#fcfaff',
  Lightgrey = '#d3d3d3',
  Violet100 = '#522699',
  Violet80 = 'rgba(82, 38, 153, 0.6)',
  Violet60 = '#dbcbf5',
  Violet40 = 'rgba(219, 203, 245, 0.6)',
  Violet20 = '#f3eefc',
  Blue = '#14aaff',
  Yellow = '#ffbb00',
  Rose = '#ff9d9d',
  LightPurple = '#a784f3',
  LightBlue = '#7fd7ff',
}